# Plugins/BTCPayServer.Plugins.SatoshiTickets/Resources/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SatoshiTickets Greenfield API",
    "description": "REST API for the SatoshiTickets BTCPay Server plugin. Provides endpoints for managing events, ticket types, tickets, orders, and check-ins.\n\nAll endpoints are scoped to a specific store and require a Greenfield API key with appropriate permissions.\n\n**Authentication:** Pass your API key in the `Authorization` header:\n```\nAuthorization: token YOUR_API_KEY\n```\n\nCreate an API key under **Account > Manage Account > API Keys** in BTCPay Server.",
    "version": "1.2.2",
    "contact": {
      "name": "SatoshiTickets",
      "url": "https://github.com/TChukwuleta/BTCPayServerPlugins"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1/stores/{storeId}/satoshi-tickets",
      "description": "Current BTCPay Server instance",
      "variables": {
        "storeId": {
          "default": "your-store-id",
          "description": "The store ID"
        }
      }
    }
  ],
...
  "paths": {
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "List events",
        "description": "Returns all events for the store (active and disabled). Use `expired=true` to get only past events.",
        "operationId": "getEvents",
        "parameters": [
          {
            "name": "expired",
            "in": "query",
            "description": "If `true`, returns only events where `startDate` is in the past.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventData"
                  }
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create event",
        "description": "Creates a new event. Events are created in `Disabled` state by default. Set `enable: true` to activate immediately.\n\nYou must create at least one ticket type before activating an event.",
        "operationId": "createEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventData"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}": {
      "get": {
        "tags": ["Events"],
        "summary": "Get event",
        "description": "Returns a single event by ID (active or disabled).",
        "operationId": "getEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Event details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      },
      "put": {
        "tags": ["Events"],
        "summary": "Update event",
        "description": "Updates an existing event.\n\n**Event Logo handling:**\n- Send `eventLogoFileId: \"file-id\"` to set/change the logo\n- Send `eventLogoFileId: \"\"` (empty string) to remove it\n- Omit the field or send `null` to leave unchanged",
        "operationId": "updateEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Delete event",
        "description": "Deletes the event along with all its ticket types and tickets.\n\n**Blocking condition:** Deletion is blocked if the event is in the future and has sold tickets. Past events can always be deleted.",
        "operationId": "deleteEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted"
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Cannot delete — event is in the future and has sold tickets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                },
                "example": {
                  "code": "event-has-active-tickets",
                  "message": "Cannot delete event as there are active ticket purchases and the event is in the future"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/toggle": {
      "put": {
        "tags": ["Events"],
        "summary": "Toggle event status",
        "description": "Flips event state between `Active` and `Disabled`. Cannot activate an event that has no ticket types.",
        "operationId": "toggleEventStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Event status toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Cannot activate — no ticket types exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                },
                "example": {
                  "code": "no-ticket-types",
                  "message": "Cannot activate event without ticket types. Create at least one ticket type first."
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/logo": {
      "post": {
        "tags": ["Event Logo"],
        "summary": "Upload event logo",
        "description": "Uploads an image file as the event logo. Only requires store-level permissions.\n\n**Constraints:**\n- File must be an image (`image/*` content type)\n- Maximum file size: 1 MB",
        "operationId": "uploadEventLogo",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The image file to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logo uploaded — returns updated event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Upload failed (not an image, too large, etc.)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                },
                "example": {
                  "code": "logo-upload-failed",
                  "message": "The uploaded file is not a valid image"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      },
      "delete": {
        "tags": ["Event Logo"],
        "summary": "Delete event logo",
        "description": "Removes the logo from the event and deletes the file from storage.",
        "operationId": "deleteEventLogo",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "responses": {
          "200": {
            "description": "Logo removed — returns updated event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "API_Key": {
        "type": "apiKey",
        "description": "BTCPay Server Greenfield API Key. Pass as `Authorization: token <key>`. Create under Account > Manage Account > API Keys. Required permission: `btcpay.store.canmodifystoresettings`.",
        "name": "Authorization",
        "in": "header"
      }
    },
    "parameters": {
      "eventId": {
        "name": "eventId",
        "in": "path",
        "description": "The event ID.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "EventNotFound": {
        "description": "Event not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GreenfieldError"
            },
            "example": {
              "code": "event-not-found",
              "message": "The event was not found"
            }
          }
        }
      }
    },
    "schemas": {
      "CreateEventRequest": {
        "type": "object",
        "required": ["title", "startDate"],
        "properties": {
          "title": {
            "type": "string",
            "description": "Event title."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Event description (HTML allowed)."
          },
          "eventType": {
            "type": "string",
            "enum": ["Physical", "Virtual"],
            "default": "Physical",
            "description": "Type of event."
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Event venue or location."
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Event start date (ISO 8601). Must be in the future."
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Event end date. Must be after `startDate`."
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "description": "3-letter currency code (e.g. `EUR`, `USD`, `BTC`). Defaults to store currency if omitted.",
            "example": "EUR"
          },
          "redirectUrl": {
            "type": "string",
            "nullable": true,
            "description": "URL to redirect customer after successful payment."
          },
          "emailSubject": {
            "type": "string",
            "nullable": true,
            "description": "Email subject template. Supports placeholders: `{{Title}}`, `{{Location}}`, `{{Name}}`, `{{Email}}`, `{{Description}}`, `{{EventDate}}`, `{{Currency}}`."
          },
          "emailBody": {
            "type": "string",
            "nullable": true,
            "description": "Email body template. Same placeholders as `emailSubject`."
          }
        }
      },
      "UpdateEventRequest": {
        "type": "object",
        "required": ["title", "startDate"],
        "properties": {
          "title": {
            "type": "string",
            "description": "Event title."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Event description (HTML allowed)."
          },
          "eventType": {
            "type": "string",
            "enum": ["Physical", "Virtual"],
            "description": "Type of event."
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Event venue or location."
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Event start date (ISO 8601)."
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Event end date. Must be after `startDate`."
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "description": "3-letter currency code."
          },
          "redirectUrl": {
            "type": "string",
            "nullable": true,
            "description": "URL to redirect customer after successful payment."
          },
          "emailSubject": {
            "type": "string",
            "nullable": true,
            "description": "Email subject template."
          },
          "emailBody": {
            "type": "string",
            "nullable": true,
            "description": "Email body template."
          }
        }
      },
      "EventData": {
        "type": "object",
        "description": "Full event representation.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Event ID."
          },
          "storeId": {
            "type": "string",
            "description": "Store ID."
          },
          "title": {
            "type": "string",
            "description": "Event title."
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Event description."
          },
          "eventType": {
            "type": "string",
            "enum": ["Physical", "Virtual"],
            "description": "Type of event."
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Event venue or location."
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Event start date."
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Event end date."
          },
          "currency": {
            "type": "string",
            "description": "Currency code.",
            "example": "EUR"
          },
          "redirectUrl": {
            "type": "string",
            "nullable": true,
            "description": "Post-payment redirect URL."
          },
          "emailSubject": {
            "type": "string",
            "nullable": true,
            "description": "Email subject template."
          },
          "emailBody": {
            "type": "string",
            "nullable": true,
            "description": "Email body template."
          },
          "eventState": {
            "type": "string",
            "enum": ["Active", "Disabled"],
            "description": "Current event state."
          },
          "purchaseLink": {
            "type": "string",
            "nullable": true,
            "description": "Public ticket purchase page URL."
          },
          "ticketsSold": {
            "type": "integer",
            "description": "Number of settled tickets sold."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the event was created."
          }
        }
      },
...
      "GreenfieldError": {
        "type": "object",
        "description": "Standard Greenfield API error.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "array",
        "description": "Validation error list.",
        "items": {
          "type": "object",
          "properties": {
            "path": {
              "type": "string",
              "description": "Field that caused the error."
            },
            "message": {
              "type": "string",
              "description": "Validation message."
            }
          }
        }
      }
    }
  }
}